refactor(extract): isolate case reporter semantics
